Add single vector R-precision evaluator and refactor FewNERD sentence evaluation
